mgproject/requirements.txt
AgentLogin==3.0.0
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
captcha==0.4
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
Django==2.2
django-crontab==0.7.1
django-haystack==3.2.1
django-redis==5.2.0
elasticsearch==2.4.1
idna==3.3
itsdangerous==1.1.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mutagen==1.45.1
mysqlclient==2.1.0
packaging==21.3
pika==1.2.1
Pillow==9.1.0
prompt-toolkit==3.0.29
py3Fdfs==2.2.0
pycparser==2.21
pycryptodomex==3.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.8
pytest-runner==6.0.0
python-alipay-sdk==3.0.4
pytz==2022.1
redis==4.2.2
requests==2.27.1
ronglian-sms-sdk==1.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
